Extract config package
